Add routing for local development
