+++ plugins/polio/js/package.json
@@ -13,7 +13,7 @@
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
         "react-query": "^3.15.2",
-        "react-router-dom": "^5.2.0",
+        "react-router": "^3.2.6",
         "react-table": "^7.7.0",
         "regenerator-runtime": "^0.13.7",
         "yup": "^0.32.9"
