+++ plugins/polio/js/src/components/Dashboard.js
@@ -15,6 +15,7 @@ import {
     Tabs,
     Typography,
 } from '@material-ui/core';
+
 import merge from 'lodash.merge';
 import EditIcon from '@material-ui/icons/Edit';
 import AddIcon from '@material-ui/icons/Add';
