+++ plugins/polio/js/src/index.html
@@ -3,7 +3,7 @@
     <head>
         <meta charset="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>React Webpack App</title>
+        <title>Polio app</title>
     </head>
 
     <body>
